k210_hal::plic
Trait PlicExt

pub trait PlicExt {
    type Interrupt: Nr;

    // Required methods
    fn is_enabled(hart_id: usize, interrupt: Self::Interrupt) -> bool;
    unsafe fn unmask(hart_id: usize, interrupt: Self::Interrupt);
    fn mask(hart_id: usize, interrupt: Self::Interrupt);
    fn get_priority(interrupt: Self::Interrupt) -> Priority;
    unsafe fn set_priority(interrupt: Self::Interrupt, prio: Priority);
    fn get_threshold(hart_id: usize) -> Priority;
    unsafe fn set_threshold(hart_id: usize, threshold: Priority);
    fn claim(hart_id: usize) -> Option<Self::Interrupt>;
    fn complete(hart_id: usize, interrupt: Self::Interrupt);
    fn is_pending(interrupt: Self::Interrupt) -> bool;
}

Extension trait for PLIC interrupt controller peripheral

Required Associated Types

type Interrupt: Nr

Interrupt wrapper type

Required Methods

fn is_enabled(hart_id: usize, interrupt: Self::Interrupt) -> bool

Is this M-Mode interrupt enabled on given hart?

unsafe fn unmask(hart_id: usize, interrupt: Self::Interrupt)

Enable an interrupt for a given hart

fn mask(hart_id: usize, interrupt: Self::Interrupt)

Disable an interrupt for a given hart

fn get_priority(interrupt: Self::Interrupt) -> Priority

Get global priority for one interrupt

unsafe fn set_priority(interrupt: Self::Interrupt, prio: Priority)

Globally set priority for one interrupt

fn get_threshold(hart_id: usize) -> Priority

Get priority threshold for a given hart

unsafe fn set_threshold(hart_id: usize, threshold: Priority)

Set the priority threshold for a given hart

fn claim(hart_id: usize) -> Option<Self::Interrupt>

Mark that given hart have claimed to handle this interrupt

fn complete(hart_id: usize, interrupt: Self::Interrupt)

Mark that given hart have completed handling this interrupt

fn is_pending(interrupt: Self::Interrupt) -> bool

Is this interrupt claimed and under procceeding?

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.